refactor: extract FullscreenMobileDismiss for fullscreen portals
